# docs/openapi.yaml
openapi: "3.0.3"
info:
  title: Outreach Service API
  version: "1.0.0"
  description: API per gestione outreach email aziende
servers:
  - url: https://outreach.example.com
    description: Produzione
  - url: http://localhost:3002
    description: Sviluppo
paths:
...
  /api/auth/login:
    post:
      summary: Invia OTP via email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email: { type: string, format: email }
      responses:
        "200":
          description: Email inviata se utente registrato
        "400":
          description: Validazione fallita
  /api/auth/verify:
    post:
      summary: Verifica OTP o magic link e restituisce JWT
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email: { type: string, format: email }
                otp: { type: string, minLength: 6, maxLength: 6 }
                token: { type: string }
      responses:
        "200":
          description: Login riuscito, cookie token impostato
        "401":
          description: Codice non valido o scaduto
...
  /api/scraper/url:
    post:
      summary: Scarica contenuto URL
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                url: { type: string, format: uri }
              required: [url]
      responses:
        "200": { description: Contenuto scaricato }
        "400": { description: URL non valido }
  /api/scraper/extract:
    post:
      summary: Estrai dati azienda da website (AI)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                website: { type: string }
                companyId: { type: integer }
              required: [website]
      responses:
        "200": { description: Dati estratti + aggiornati su DB se companyId fornito }
        "400": { description: Validazione fallita }
  /api/scraper/generate-email:
    post:
      summary: Genera bozza email per azienda via AI
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                companyId: { type: integer }
              required: [companyId]
      responses:
        "200": { description: Bozza generata e salvata }
        "422": { description: Blacklist o validazione fallita }
...
  /dashboard:
    get:
      summary: Dashboard marketing con KPI, funnel, grafici
      responses:
        "200": { description: Pagina HTML dashboard }
        "302": { description: Redirect a /login se non autenticato }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: token
security:
  - cookieAuth: []